Campaign page displays player information in a usable format
Current players used to be displayed by id number and their commander's
id number.

Now displays username and commander names.

This is accomplished by loading the information into state during the
construction of the react element.

Attempts were made to use componentDidMount() however the information
was still in transit from the DB at render time so it rendered an empty
space.  In retrospect typing this out that seems pretty obvious that it
would work that way...

Also refactored the addSkill route.  Used to pass information via the
URI and using req.params.  Realized earlier that I was using a form
submit button and can pass the info via hidden fields and have a cleaner
server method.