Hotfix for Create Commander link not working in deployment
Converted from a link to a form button.  Similar elements have been
working in deployment.  This will also give visual distintion between a
commander and a lack of one.

Also added error handling for trying to start a campaign of the same
name as an existing one.   Previous behaviour just had you rolling into
the already existing campaign, which is not intended behaviour.